backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
opencv-python-headless==4.10.0.84
face-recognition==1.3.0
numpy==1.26.4
pydantic-settings==2.9.0